# backend/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/files/upload:
    post:
      tags:
      - upload-controller
      operationId: upload
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      requestBody:
        description: ZIP archive to ingest
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseUploadSessionResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/files/sessions:
    get:
      tags:
      - upload-controller
      operationId: getAvailableSessions
      parameters:
      - name: X-User-Id
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseListSessionResponseDTO"
...
components:
  schemas:
    ErrorResponse:
      type: object
      description: Unified error body. The code mirrors the HTTP status.
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          nullable: true
...
    ApiResponseListSessionResponseDTO:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/SessionResponseDTO"
...
    ApiResponseUploadSessionResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/UploadSessionResponse"
...
    UploadSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
...
    SessionResponseDTO:
      type: object
      properties:
        uuid:
          type: string
        jsonFilename:
          type: string
        uploadedAt:
          type: string